Move parse_encoded_pointer etc into cfi module
